Added basic Letter detection
Letter detection for a trie is to see if the trie already contains a sequence of letters in the correct order to form part or an entire word. Next step of letter detection is to move two two letter detection meaning that the words test and tested would share the first four nodes in the trie.